Fix #44, instantiate properly sized and aligned buffers
For unit unit tests that invoke CF PDU processing functions
on either input or output, ensure that the locally instantiated
"dummy" PDU is both sized sufficiently and aligned correctly.

This removes quite a bit of questionable casting between the
buffer types, and fixes a number of stack-smashing issues.

For completeness, this also clears (memset to 0) all instantiated
buffers, before setting values in the test.  This ensures that the
entire message structure has predictable/repeatable content.