Fix #43, correct declaration of CF_CFDP_PlaybackDir_
This helper function was declared with fixed-length string arguments
but in practice is used with other buffer lengths.  These arguments
are changed to (more conventional) const char * instead.

Note that other cases where fixed-length string arguments are used
are not changed by this commit, since it appears those are always
used with a buffer of that length.