Fix #41, no packing of CMD/TLM structs
The CMD/TLM structures are application-defined.  As such, they should
not need to depend on compiler-specific extensions like packing.

Note that the definitions (mostly) already included explicit spare bytes
where padding was likely to be needed, so the "packed" attribute was
largely unnecessary here, only serving to make the definitions NOT compilant
with C99.