Update #113, replace some UT random values with fixed values
The use of random numbers in a couple test cases was causing an
unpredictable flow through the code.  This replaces with fixed
values so get consistent flows.

Specifically, the "acknak_count" should always be less than the
"ack_limit".  The UT use of random numbers allowed it to be
greater.  However, if it is greater than the limit, it should
still follow the limit/error path, not the successful path.