Fix #111, make dispatch tables and functions consistent
Standardize the dynamic handler functions to two basic types,
one that accepts a PDU (recv) and one that does not (send).

Also create several dispatch table types, one based on
file directive code, one based on Tx sub state, and one based
on Rx sub state.

Change the dispatcher functions to use these common types
and create new dispatcher functions where there was not
a separate function already (this makes the pattern consistent).

Make all "receive" helper functions accept a pointer to the
recieved PDU and actually use that pointer to read the data.  This
substantially reduces reliance on the global and fixes some
cases where a pointer was actually passed into a function, but
ignored.  This takes a significant step toward removing the
global entirely, but does not do so yet.