Update #116, cleanup unit tests
A significant cleanup and overhaul of CF unit tests
to follow patterns that are more consistent with
other CFE/OSAL test modules.

Since the FSW change in this PR requires significant test updates
to go with it, this replaces all the "cfdp" tests with new
implementations where there is a 1:1 ratio between test
functions and implementation functions.

Ths also does some minor cleanup on the FSW side where needed
for testability.

Note that all stub files in this version are now directly generated
using the tool provided with UtAssert.  These generated stubs should
not be modified.