Fix #32, compile out CF_Asserts by default
Changes CF_Assert to be opt-in rather than opt-out, so that under normal
verification and validation the asserts will _not_ be included, but they
can still be added back during development, if desired.

They mainly exist as notes to developers as to what is supposed to be true,
once debugged it is impossible to get these conditions, by definition.

Also removes one channel calculation that was only for assert.  Note that
the same condition is asserted later, so it was redundant anyway.