Fix #20, major doxygen documentation update
Make sure _all_ functions in CF have a doxygen block on the declaration
describing the inputs and outputs and a general summary of what the function does.

Additionally describe the purpose of structures and structure members where
it is not totally obvious from the name.

Note: this also renames one member, CF_max_chunks, that appears to have been
improperly scooped up in a previous batch rename.