Fix #325, add more concise transaction status code
Adds a new concept of "transaction status" to replace use of CFDP
condition code to indicate the result of a transaction.  To aid in
transition this is equivalent in numeric value to the defined CFDP CC
values but is extended with additional values for other conditions
that can occur in the implementation but do not necessarily result
in sending a FIN/EOF PDU.

This also adds setting of Transaction Status for some off-nominal
cases where no CFDP CC was set.