Fix #184, work around throttle sem creation race
Adds a retry loop around OS_CountSemGetIdByName, because if this sem is
created by another app there may be some delay until the other app gets
to the point where it creates the sem.  This works around the race
condition.

A retry limit is also imposed so CF will not spin here forever.