Fix #418, reorganize header files and restructure messages
Move the header files to a "config" directory as documented in the CFE
guidelines, which allows easier override.

Organize the content to adhere to the patterns.

Specifically - adds a dedicated structure type for each cmd, and also
puts the payload into a separate sub-structure.

Command handlers and tests updated accordingly (many of which were simplified).