Fix #435, simplify state machines and add holdover
Updates the state machines to more reliably handle the exception cases
and atypical transitions between states.

Importantly:
 - There are now separate "undef" and "init" states for free transactions
   and newly-allocated transactions, respectively.  This replaces a
   single "idle" state and makes these distinguishable.
 - Adds a holdover state at the end of a transaction and consolidates
   closeout synchronization into this state.  Instead of a single reset at
   the end that wipes everything and immediately returns the transaction
   to the free list, the process of freeing a transaction now has two
   parts and the final return to the free list is done by the tick
   processor after the holdover period.

The segfault issue described in the ticket was the result of these two
areas of weakness in the code working together to make an extra bad
result.  The sequence was a situation where a brand-new transaction was
started and then immediately reset/closed within the same PDU processing
cycle.  Because (a) the transaction was completely wiped out and (b) the
wiped out transaction was indistinguishable from the newly-allocated
transaction, and (c) some atypical state machine transitions occurred
causing it to effectively go through some code twice in RecvIdle.

Utimately because the second exit from RecvIdle cannot tell the
difference between a txn that is already freed vs one that is left in
its original state and still needed to be cleaned, it tried to clean it
twice (basically like a double-free bug).

This addresses the fundamental problem on both sides by first making
brand-new transactions distinguishable from free transactions (so this
alone would have avoided the segfault) but also deferring the final
"free" until we can guarantee that nothing else will use the transaction
struct (the holdover).  Now there is nothing in any PDU processor that
will put the transaction into a state where it can no longer be safely
accessed, so even if there are other atypical jumps around the state
machine, it will not cause problems (or at least not a segfault) if
something finished the transaction in the middle.  The actual free is
now done at the end of tick processing which is the very last thing
to occur on the wakeup cycle.