Fix #481, simplify and consolidate state machines
Rework state machines such that all state changes are performed in the
tick processor, rather than scattered around in various parts of PDU
processing.

This removes many dependencies and assumptions about the order of
arrival of PDUs and makes the protocol handling more robust and accurate
per the blue book spec.

Notably, this fixes a number of important concerns:
 - Out of order MD arrival is no longer an exception, the state machine
   is identical regardless if MD comes first or not.
 - Out of order arrival is handled in class 1 (PDUs can arrive in any
   order)
 - Logic for file retention is consistently applied in both success and
   error cases - always goes through this part of the state machine.
 - Allows transferring a zero byte file (previously would error out
   because of the way the "read" call was reached).
 - Files are no longer updated in place.  This means an existing file
   will not be clobbered until a complete new file is received.
 - Bit errors in an FD PDU header simply drop the PDU (it will be NAKed)
   rather than aborting the entire transaction.
 - ACK/NAK counting is accurately implemented in class 2.
 - The CC response in EOF/FIN packets will be accurate when the ACK
   limit is reached.
 - Tx-side tick processing simplified, there is now just one tick
   processor, all the special "cont" and "early_exit" flags removed.

The bottom line: all transactions will proceed through the state machine
in the exact same way, whether class 1 or class 2, with transitions all
handled in the same place and in a consistent manner, going through all
the same checks and validations on every transition in every
transaction, regardless of PDU arrival order.  This greatly improves
adherence to the CFDP protocol spec and ability to deal with imperfect
PDU delivery.

Bonus item:  The "closure request" feature of class 1 transfers now is
supported and works in the engine.  (Only missing part is a way to
request it when initiating a Tx transaction)